change home and block page intetion link
