+++ packages/site/src/pages/Blocks/BlocksList.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useMemo } from "react";
 
-import { Spinner, Table, Link, AddressLink, DateShow, Number } from "../../components";
+import { Spinner, Table, Link, DateShow, Number, AddressLinkExtend } from "../../components";
 import { useRedux } from "../../shared";
 import api from "../../services/api";
 import TableService from "../../services/tableService";
@@ -31,7 +31,7 @@ export default function BlocksList() {
             number: <Link parent="blocks" value={data.number} />,
             hash: <Link parent="blocks" style={{ width: 138 }} className="text-truncate" value={data.hash} />,
             time: <DateShow value={data.time} />,
-            producer: <AddressLink isValidator value={data.producer} />,
+            producer: <AddressLinkExtend value={data.producer} />,
             extrinsics: <Number value={data.extrinsics} />,
             eventCount: <Number value={data.event_count} />
           };
